Draw keyboard using pygame primitives
Initially the keyboard was made by a set of images with
all combinations of highlighed keys. This way turns a
changing in the keyboard layout much more time consuming.

We're now drawing the keyboard using `rect()` and `font`
elements from pygame. The text displayed in the keyboard
keys are controlled by a settings file.