Fix sound for the new keyboard layout
The drawing of keyboard changes the dimensions of the keyboard
which leads to an IndexError on trying to load a sound for a
row that doesn't exist.

Initially we were loading sounds based on row and col indexes
which may vary in case of a keyboard layout. In order to uncouple
these we will now pass to load sounds based on the number and
letter names.